Create the foundation of website style using CSS
The purpose of this commit is to add simple, foundational elements of html and CSS code in order to begin styling the visuals of the main webpage. There is now a functioning CSS file titled main.css which includes barebones CSS elements with varying changes to background color, font color, font size, and font type. There are also barebones html elements to serve as sample text.

Almost all of the properties in the CSS file are meant to be temporary. They are all test properties made without consideration to the final design.