Reorganized nameplate style.
The new nameplate style was created by heavily adjusting the code in main.css. What was just a child variation of div is now called "nameplate". The nameplate class also has two new child classes. The "title" class is for the title of the displayed video game. The "description" class is for the description... or 'final words' of said video game.

This overall organization of the nameplate is mostly how I envision the final product to look, however there are a vast multitude of adjustments that are planned for the finer details such as color, padding size, text size, font, etc.

Alternate styles for differing viewports will also be added later in development.