Infinite Scrolling using React and AJAX
I converted this project into a React app using create-react-app and node.js.
I'm only using React functionality to implement infinite scrolling. This is accomplished
using IntersectionObserver to detect when an element is on the screen.
IntersectionObserver plus an updated version of our code used to generate nameplates
is used to create the effect of Infinite Scrolling.

The file structure has changed as well. As a result of installing React,
node_modules, public, and src are all new folders. The main.js file
has been deleted, but most of the code has been moved and adapted
to the new file called app.js. index.js is another new JS file created by
create-react-app, but app.js is where the real meat of the project is.
I also moved the covers folder containing image files into src.

Ever since I started using a development server, the images for the
nameplates haven't been showing. I believe that merging with the
master branch will solve the issue, but there are other options
such as uploading the images to an internet image host.

There is likely more work to be done in order to get infinite scrolling
just right. The code where I use an AJAX call to access my json file can
probably be put into a seperate function whereas it currently sits in the
main app() function. This is a potential improvement I intend to look into
before finalizing the project.

A significant change has been made in regards to the structure of the webpage.
We are no longer using the graveyard->column->nameplate structure to display
the nameplates for each game. Columns have been removed and the new structure
is now graveyard->nameplates. The reason for this change was to ease the use
of IntersectionObserver and to create more flexibility with the page elements
as the project comes closer to a finish. Main.css has been updated to retain
the column alignments of the nameplates without needing the column div.

Next up is using our infinite scrolling code to implement sorting by console.
I will keep an eye on how the images behave on the master branch as well.