Bug fixes, style adjustment, and code clean-up.
nameplates.json:
- Added missing/incorrect images.
- Added games that I missed.

index.html:
- Reorganized <p> element to be below #toolbar.
- <p> text now describes the purpose of the website.
- Removed PS3 and PS4 buttons because there are no PS3 or PS4 games in the data.

main.css:
- Touched up various elements of the style to be as readable and pleasant
as possible to view in as many window sizes as possible.
- Reorganized the order of the style rules for easier readability (doesn't affect functionality.

README.md:
- Updated the readme to more accurately represent the completed project.

Various small alterations were made to the code comments throughout the javascript files.

This being my first personal programming project, I'm not sure if the following has a place
in a commit message, but I feel it's necessary to say:
- This project was made by Ryan Tommy Tavares.
- email: RyanTommyTavares@gmail.com